[FLINK-37481][table] Split test into multiple files
